LayoutTests/imported/w3c/web-platform-tests/intersection-observer/observer-callback-arguments-expected.txt


PASS Callback is invoked with |this| value of IntersectionObserver and two arguments
